Switched the component union in cpu.zig
to an extern union. Apparently in Zig 0.13 and beyond, you can't
arbitrarily change whichever type in the union is active while retaining
it's contents without ptrCast'ing the shit out of everything. Ugly as
fuck and I want to meet whoever thought that was a good idea. This
worked just as it was in 0.12 and before.