Switched range-based switch statements to ifs
Range-based switch statements required comptime-known values

This complicated memory mapping, as the limits had to be known at
compile-time, preventing configuration post-compilation.

This also complicated the function signatures of RV32 cores, since I
needed to use generics in order to take parameters and basically
pseudo-template my bus.
In order to instantiate a bus to *pass* to the core, i.e. Bus(256){}, I
also needed to taint the function signature by putting Bus(256) as the
type for the Bus parameter.

None of these issues exist with the If based implementation, and is,
in-fact, moire faithful to a bus architecture.